requirements/development.txt
- r production.txt

factory-boy==3.3.0
pre-commit==3.5.0
ruff==0.1.5
pytest-django==4.7.0